@@ -1,2 +1,2 @@
-export * from "./auth";
 export * from "./projects";
+export * from "./auth";
