@@ -2,7 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { api } from "../..";
 
 const getSession = async () => {
-  const res = await api.get("/auth/session", {});
+  const res = await api.get("/web/auth/session", {});
   console.log(res.data);
   return res.data;
 };
