@@ -1,5 +1,5 @@
 export function usePath(): string {
-  return new URL(window.location.href).hash;
+  return new URL(window.location.href).pathname;
 }
 
 export { useHomePage, useSettingsPage, useProfilePage } from "./pages";
